Add compatibility with Gradle 4.10 and above
This commits adds compatibility for Gradle 4.10 onward (tested on 4.10
and 5,4), while retainting compatibiliy with 4.9

Due to the ABI-incompatible changes in some internal gradle classes, we
need to use bytecode manipulation to ensure that our built class files
are compatible with both Gradle 4.9 and Gradle >= 4.10

Bytecode manipulation is performed at built time. A custom Gradle task
is used to read in the compiled class files from disk, and write out a
modified version to the final jar artifact.

In order to make as few bytecode modifications as possible, this commit
bumps the Gradle wrapper version to 4.10. This means that we're
compiling against Gradle 4.10, and using bytecode manipulation to retain
compatibility with 4.9. Doing the reverse (compiling against 4.9) would
be significantly more difficult, as we would need to statically
reference classes that exist in Gradle 4.10 but not 4.9 (specifically,
RepositoryDescriptor)

We perform two different bytecode transformations:

1. We modify the call to 'super()' in GradleRepositoryAdapter. In Gradle
4.10, the suepr constructor takes one argument, but in 4.9, it takes
zero arguments. In order to allow GradleRepositoryAdapter to compile
normally, we write a 'fake' call to 'super(null)', and replace with a
call to 'super()'.

2. We delete the method 'getDescriptor' from GradleRepositoryAdapter. In
Gradle 4.9, its return type does not exist, and will cause a
NoClassDefFoundError when Gradle attempts to classload it via
Class#getDeclaredMethods. However, it's necessary to include
'getDescriptor' so that GradleRepositoryAdapter (we need to override the
abstract method in a parent class).