Avoid modifying final fields in favor of making new objects
Because of Java 16's strong encapsulation of JDK internals, the current
trick for modifying final fields via reflection into Field#modifiers
cannot work without additional command-line arguments to the JVM.

To avoid reflecting into the JDK, we have to avoid reflection to
edit final fields. So, instead of modifying fields of existing objects,
we create new objects using most of the objects from the old object,
except substituting our own objects as needed.

Fortunately, Gradle doesn't do much activity in the constructors of
its objects. This means that it may be _possible_ to create these
objects and substitute them without causing issues due t
misconfigurations. To be absolutely sure of this would require analysis
of the whole code path leading up to the BaseRepositoryFactory. As
this is _very_ time consuming, a better endeavor would be testing the
code live and fixing errors as they popup (with stacktrace information).